database: Add GetContinuationToken to DBClientIterator
A new function NewQueryItemsSinglePageIterator alters the behavior
of QueryItemsIterator to stop after the first "page" of items and
record the Cosmos DB continuation token. The continuation token
can be retrieved from the iterator with GetContinuationToken.

A QueryItemsIterator created with NewQueryItemsIterator will never
have a continuation token because it iterates until the last item.

The in-memory cache also adds a GetContinuationToken method to its
iterator implementation to fulfill the interface contract, but it
always returns an empty string.