frontend: Rename StartOperation to ExposeOperation
The operation item must now be created in the database prior to
calling ExposeOperation. ExposeOperation does all its processing
in a database update callback.

This is because there is an increasing number of cases where we
create an implicit async operation with no visible status endpoint.
Calling ExposeOperation makes an implicit async operation explicit,
with a status endpoint for ARM to poll. Hence the rename.

The tradeoff is explicit asyncrhonous operations now require two
database operations (create and update) but it helps make the RP
logic cleaner. This could possibly be mitigated in the future by
using Cosmos DB's transactional batch operations, but it's gonna
take some serious refactoring to get there.